Generate Long parameters for ordinal arguments
Special case Long.MAX_VALUE ordinal formatting test on APIs < 26: ICU versions on older Android platforms lose bits by internally converting Long to Double

+++ plugin/src/main/java/app/cash/paraphrase/plugin/ArgumentTypeResolver.kt
@@ -65,7 +65,7 @@ internal fun resolveArgumentType(tokenTypes: List<TokenType>): KClass<*>? =
     DateTimeWithZone -> ZonedDateTime::class
     Offset -> ZoneOffset::class
     Duration -> KotlinDuration::class
-    Ordinal, SelectOrdinal -> Int::class
+    Ordinal, SelectOrdinal -> Long::class
     Select -> String::class
     NoArg -> Nothing::class
   }
